#167 - Move classes in package 'de.tudarmstadt.ukp.wikipedia.api.tutorial' to 'tutorial' module
- Moves classes T1a/b/c_... to T6_... that existed in package '...api.tutorial' to the tutorial module
