--- de.tudarmstadt.ukp.wikipedia.api/src/main/java/de/tudarmstadt/ukp/wikipedia/api/tutorial/T1a_HelloWorld.java
@@ -1,57 +1,0 @@
-/*******************************************************************************
- * Copyright 2017
- * Ubiquitous Knowledge Processing (UKP) Lab
- * Technische Universität Darmstadt
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *******************************************************************************/
-package de.tudarmstadt.ukp.wikipedia.api.tutorial;
-
-import de.tudarmstadt.ukp.wikipedia.api.DatabaseConfiguration;
-import de.tudarmstadt.ukp.wikipedia.api.Page;
-import de.tudarmstadt.ukp.wikipedia.api.WikiConstants;
-import de.tudarmstadt.ukp.wikipedia.api.Wikipedia;
-import de.tudarmstadt.ukp.wikipedia.api.exception.WikiApiException;
-
-/**
- * Tutorial 1a
- *
- * Get the text of a wikipedia article.
- * The text will be formatted with MediaWiki markup.
- *
- * Throws an exception, if no page with the given title exists.
- *
- *
- */
-public class T1a_HelloWorld implements WikiConstants {
-
-    public static void main(String[] args) throws WikiApiException {
-
-        // configure the database connection parameters
-        DatabaseConfiguration dbConfig = new DatabaseConfiguration();
-        dbConfig.setHost("SERVER_URL");
-        dbConfig.setDatabase("DATABASE");
-        dbConfig.setUser("USER");
-        dbConfig.setPassword("PASSWORD");
-        dbConfig.setLanguage(Language.german);
-
-        // Create a new German wikipedia.
-        Wikipedia wiki = new Wikipedia(dbConfig);
-
-        // Get the page with title "Hello world".
-        // May throw an exception, if the page does not exist.
-        Page page = wiki.getPage("Hello world");
-        System.out.println(page.getText());
-
-    }
-}
