+++ de.tudarmstadt.ukp.wikipedia.revisionmachine/src/test/resources/db/wikiapi_simple_20090119_stripped.properties
@@ -1,5 +1,5 @@
 #HSQL Database Engine 2.4.0
-#Tue Jul 03 15:03:17 CEST 2018
+#Fri Jul 06 17:23:13 CEST 2018
 version=2.4.0
 modified=yes
 tx_timestamp=0
