+++ de.tudarmstadt.ukp.wikipedia.tutorial/src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/api/T1a_HelloWorld.java
@@ -15,32 +15,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package de.tudarmstadt.ukp.wikipedia.api.tutorial;
+package de.tudarmstadt.ukp.wikipedia.tutorial.api;
 
-import java.util.Set;
-import java.util.TreeSet;
-
-import de.tudarmstadt.ukp.wikipedia.api.Category;
 import de.tudarmstadt.ukp.wikipedia.api.DatabaseConfiguration;
 import de.tudarmstadt.ukp.wikipedia.api.Page;
 import de.tudarmstadt.ukp.wikipedia.api.WikiConstants;
 import de.tudarmstadt.ukp.wikipedia.api.Wikipedia;
 import de.tudarmstadt.ukp.wikipedia.api.exception.WikiApiException;
-import de.tudarmstadt.ukp.wikipedia.api.exception.WikiPageNotFoundException;
-
 
 /**
- * Tutorial 5
+ * Tutorial 1a
  *
- * Wikipedia categories are used as a kind of semantic tag for pages.
- * They are organized in a thesaurus like structure.
+ * Get the text of a wikipedia article.
+ * The text will be formatted with MediaWiki markup.
  *
- * If we get all pages assigned to categories in the sub-tree under the category for "Towns in Germany",
- *   we can get a quite long list of towns in Germany.
+ * Throws an exception, if no page with the given title exists.
  *
  *
  */
-public class T5_TownList implements WikiConstants {
+public class T1a_HelloWorld implements WikiConstants {
 
     public static void main(String[] args) throws WikiApiException {
 
@@ -55,33 +48,10 @@ public class T5_TownList implements WikiConstants {
         // Create a new German wikipedia.
         Wikipedia wiki = new Wikipedia(dbConfig);
 
-        // Get the category "Towns in Germany"
-        String title = "Towns in Germany";
-        Category topCat;
-        try {
-            topCat = wiki.getCategory(title);
-        } catch (WikiPageNotFoundException e) {
-            throw new WikiApiException("Category " + title + " does not exist");
-        }
-
-        // Add the pages categorized under "Towns in Germany".
-        Set<String> towns = new TreeSet<String>();
-        for (Page p : topCat.getArticles()) {
-            towns.add(p.getTitle().getPlainTitle());
-        }
-
-        // Get the pages categorized under each subcategory of "Towns in Germany".
-        for (Category townCategory : topCat.getDescendants()) {
-            for (Page p : townCategory.getArticles()) {
-                towns.add(p.getTitle().getPlainTitle());
-            }
-            System.out.println("Number of towns: " + towns.size());
-        }
-
-        // Output the pages
-        for (String town : towns) {
-            System.out.println(town);
-        }
+        // Get the page with title "Hello world".
+        // May throw an exception, if the page does not exist.
+        Page page = wiki.getPage("Hello world");
+        System.out.println(page.getText());
 
     }
 }
