+++ de.tudarmstadt.ukp.wikipedia.tutorial/src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/api/T2_PageInfo.java
@@ -15,25 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package de.tudarmstadt.ukp.wikipedia.api.tutorial;
+package de.tudarmstadt.ukp.wikipedia.tutorial.api;
 
-import de.tudarmstadt.ukp.wikipedia.api.Category;
 import de.tudarmstadt.ukp.wikipedia.api.DatabaseConfiguration;
 import de.tudarmstadt.ukp.wikipedia.api.Page;
-import de.tudarmstadt.ukp.wikipedia.api.Title;
 import de.tudarmstadt.ukp.wikipedia.api.WikiConstants;
 import de.tudarmstadt.ukp.wikipedia.api.Wikipedia;
 import de.tudarmstadt.ukp.wikipedia.api.exception.WikiApiException;
 import de.tudarmstadt.ukp.wikipedia.api.exception.WikiPageNotFoundException;
 
+
 /**
- * Tutorial 3
+ * Tutorial 2
  *
- * Even more things to do with a Wikipedia page.
+ * A page provides a number of informative methods.
  *
  *
  */
-public class T3_PageDetails implements WikiConstants {
+public class T2_PageInfo implements WikiConstants {
 
     public static void main(String[] args) throws WikiApiException {
 
@@ -45,7 +44,7 @@ public class T3_PageDetails implements WikiConstants {
         dbConfig.setPassword("PASSWORD");
         dbConfig.setLanguage(Language.german);
 
-        // Create a new German wikipedia.
+        // Create a new German wikipedia
         Wikipedia wiki = new Wikipedia(dbConfig);
 
         String title = "Hello world";
@@ -56,40 +55,25 @@ public class T3_PageDetails implements WikiConstants {
             throw new WikiApiException("Page " + title + " does not exist");
         }
 
-        StringBuilder sb = new StringBuilder();
-
         // the title of the page
-        sb.append("Queried string : " + title + LF);
-        sb.append("Title          : " + page.getTitle() + LF);
-        sb.append(LF);
+        System.out.println("Queried string       : " + title);
+        System.out.println("Title                : " + page.getTitle());
 
-        // output the page's redirects
-        sb.append("Redirects" + LF);
-        for (String redirect : page.getRedirects()) {
-            sb.append("  " + new Title(redirect).getPlainTitle() + LF);
-        }
-        sb.append(LF);
+        // whether the page is a disambiguation page
+        System.out.println("IsDisambiguationPage : " + page.isDisambiguation());
 
-        // output the page's categories
-        sb.append("Categories" + LF);
-        for (Category category : page.getCategories()) {
-            sb.append("  " + category.getTitle() + LF);
-        }
-        sb.append(LF);
+        // whether the page is a redirect
+        // If a page is a redirect, we can use it like a normal page.
+        // The other infos in this example are transparently served by the page that the redirect points to.
+        System.out.println("redirect page query  : " + page.isRedirect());
 
-        // output the ingoing links
-        sb.append("In-Links" + LF);
-        for (Page inLinkPage : page.getInlinks()) {
-            sb.append("  " + inLinkPage.getTitle() + LF);
-        }
-        sb.append(LF);
+        // the number of links pointing to this page
+        System.out.println("# of ingoing links   : " + page.getNumberOfInlinks());
 
-        // output the outgoing links
-        sb.append("Out-Links" + LF);
-        for (Page outLinkPage : page.getOutlinks()) {
-            sb.append("  " + outLinkPage.getTitle() + LF);
-        }
+        // the number of links in this page pointing to other pages
+        System.out.println("# of outgoing links  : " + page.getNumberOfOutlinks());
 
-        System.out.println(sb);
+        // the number of categories that are assigned to this page
+        System.out.println("# of categories      : " + page.getNumberOfCategories());
     }
 }
