+++ de.tudarmstadt.ukp.wikipedia.tutorial/src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/parser/T4_InterfacingWithWikipedia.java
@@ -1,0 +1,64 @@
+/*******************************************************************************
+ * Copyright 2017
+ * Ubiquitous Knowledge Processing (UKP) Lab
+ * Technische Universität Darmstadt
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *******************************************************************************/
+package de.tudarmstadt.ukp.wikipedia.tutorial.parser;
+
+import java.util.List;
+
+import de.tudarmstadt.ukp.wikipedia.api.DatabaseConfiguration;
+import de.tudarmstadt.ukp.wikipedia.api.Page;
+import de.tudarmstadt.ukp.wikipedia.api.WikiConstants.Language;
+import de.tudarmstadt.ukp.wikipedia.api.Wikipedia;
+import de.tudarmstadt.ukp.wikipedia.api.exception.WikiApiException;
+import de.tudarmstadt.ukp.wikipedia.parser.ParsedPage;
+import de.tudarmstadt.ukp.wikipedia.parser.Section;
+import de.tudarmstadt.ukp.wikipedia.parser.mediawiki.MediaWikiParser;
+import de.tudarmstadt.ukp.wikipedia.parser.mediawiki.MediaWikiParserFactory;
+
+/**
+ * Displays the titles of the sections found in the page <i>Dog</i>.<br>
+ */
+public class T4_InterfacingWithWikipedia {
+
+	public static void main(String[] args) throws WikiApiException {
+		//db connection settings
+		DatabaseConfiguration dbConfig = new DatabaseConfiguration();
+        dbConfig.setDatabase("DATABASE");
+        dbConfig.setHost("HOST");
+        dbConfig.setUser("USER");
+        dbConfig.setPassword("PASSWORD");
+        dbConfig.setLanguage(Language.english);
+		
+		//initialize a wiki
+		Wikipedia wiki = new Wikipedia(dbConfig);
+		
+		//get the page 'Dog'
+		Page p = wiki.getPage("Dog");
+		
+		//get a ParsedPage object
+		MediaWikiParserFactory pf = new MediaWikiParserFactory();
+		MediaWikiParser parser = pf.createParser();
+		ParsedPage pp = parser.parse(p.getText());
+	
+		//get the sections of the page
+		List<Section> sections = pp.getSections();
+		
+		for(Section section : sections) {
+            System.out.println(section.getTitle());
+        }
+	}
+}
