+++ pom.xml
@@ -319,7 +319,7 @@
                     <exclude>src/main/resources/**/*</exclude>
                     <exclude>src/test/resources/**/*</exclude>
                     <exclude>.activate_rat-check</exclude>
-                    <exclude>src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/DarmstadtWikipediaArticle.txt</exclude>
+                    <exclude>src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/parser/DarmstadtWikipediaArticle.txt</exclude>
                   </excludes>
                 </configuration>
               </execution>
